+++ factory/int_intdiv.cc
@@ -65,7 +65,7 @@ InternalInteger::dividecoeff ( InternalCF * c, bool invert )
     ASSERT( invert || imm2int( c ) != 0,
             "math error: divide by zero" );
 
-    int intC = imm2int( c );
+    long intC = imm2int( c );
 
     if ( cf_glob_switches.isOn( SW_RATIONAL ) ) {
         mpz_t n, d;
